Polishing in pybind11.h: comments, moving helper functions out.
